+++ pkg/sdkclient/sink/clienttest/clienttest.go
@@ -3,10 +3,10 @@ package clienttest
 import (
 	"context"
 	"fmt"
+	sinksdk "github.com/numaproj/numaflow/pkg/sdkclient/sink/client"
 
 	sinkpb "github.com/numaproj/numaflow-go/pkg/apis/proto/sink/v1"
 	"github.com/numaproj/numaflow-go/pkg/apis/proto/sink/v1/sinkmock"
-	sinksdk "github.com/numaproj/numaflow/pkg/sink-client/client"
 	"google.golang.org/protobuf/types/known/emptypb"
 )
