+++ pkg/sinks/udsink/udsink_grpc_test.go
@@ -19,13 +19,13 @@ package udsink
 import (
 	"context"
 	"fmt"
+	"github.com/numaproj/numaflow/pkg/sdkclient/sink/clienttest"
 	"testing"
 	"time"
 
 	"github.com/golang/mock/gomock"
 	sinkpb "github.com/numaproj/numaflow-go/pkg/apis/proto/sink/v1"
 	"github.com/numaproj/numaflow-go/pkg/apis/proto/sink/v1/sinkmock"
-	"github.com/numaproj/numaflow/pkg/sink-client/clienttest"
 	"github.com/stretchr/testify/assert"
 	"google.golang.org/protobuf/types/known/timestamppb"
 )
